Fix bug where email tags got lost on move (#5434)
With disabled messages_cache and disabled kolab_notes plugin
Message-Id header is not FETCHed from the IMAP server.
So, in such a case member URI in the tag object will contain
message-id=mid:XXX, which causes problem when the message
is moved to another folder, because it cannot be found by this invalid message-id.