Allow accept/decline/delegate (again) on event update request (Bifrost#T27859)
When handling an iTip request for an updated event where
start or end date has been changed, user will be able to change
his participant status.

In such cases we display RSVP buttons instead of "Update in my calendar"
button and highlight the event date.

So, e.g. if you accepted the event and then its time has been
changed by the organizer, you will be able to decline it or vice-versa.