Add Elastic style for folders list extensions in Kolab plugins, e.g. kolab_addressbook
Libkolab plugin will contain Elastic style for all (or most) Kolab plugins
as we want to limit number of included stylesheets (and unify).

We're using less, so to build css file we require skins/elastic folder
in an appropriate place, i.e. aside of the plugins folder of this repository.